refactor: Dropdown to deprecated, implements warning FE-2007
