+++ src/__deprecated__/components/dropdown/dropdown.scss
@@ -1,8 +1,8 @@
-@import "./../../style-config/colors";
-@import "./../../style-config/z-index";
-@import "./../../style-config/general";
-@import "./../../style-config/input-field";
-@import "./../../style-config/mixins";
+@import "./../../../style-config/colors";
+@import "./../../../style-config/z-index";
+@import "./../../../style-config/general";
+@import "./../../../style-config/input-field";
+@import "./../../../style-config/mixins";
 
 $dropdown-input-icon-width: 20px !default;
