+++ src/components/dialog/dialog.stories.js
@@ -13,7 +13,7 @@ import Textbox from '../../__deprecated__/components/textbox';
 import Button from '../button';
 import Modal from '../modal';
 import DateInput from '../date';
-import Dropdown from '../dropdown/dropdown';
+import Dropdown from '../../__deprecated__/components/dropdown';
 import Checkbox from '../checkbox/checkbox';
 import getDocGenInfo from '../../utils/helpers/docgen-info';
